Split the big ugly HTML file
[FIX] Ok, one big ugly HTML file was not at all cool! So now each part
of the app has it's own file :)
[FIX] Added "http://" to the CDN's! Without them the project couldn't be
loaded from any folder and I do not want to oblige the users to install
a server to run startr. That's not the point... (unfotunatelly you'll
still need an internet connection to use startr... I'll fix this in the
near future - added to the todo list)
[FIX] Changed my GitHub username, so I updated the project's link

NOTE : from now on the startr project will use SASS, but for the moment
the .scss file has the current ugly CSS code - I'll fix this asap. I
also commited (and I always will) the .css file because this way any
user can start using the project right away